@@ -10,9 +10,9 @@
     "@testing-library/user-event": "^7.2.1",
     "graphql": "^14.6.0",
     "jwt-decode": "^2.2.0",
-    "react": "^16.13.1",
-    "react-dom": "^16.13.1",
-    "react-router-dom": "^5.1.2",
+    "react": "^16.14.0",
+    "react-dom": "^16.14.0",
+    "react-router-dom": "^5.3.2",
     "react-scripts": "3.4.1"
   },
   "scripts": {
